+++ src/Products/PloneMeeting/events.py
@@ -2,6 +2,9 @@
 
 from AccessControl import Unauthorized
 from collections import OrderedDict
+from collective.behavior.internalnumber.browser.settings import _internal_number_is_used
+from collective.behavior.internalnumber.browser.settings import decrement_if_last_nb
+from collective.behavior.internalnumber.browser.settings import decrement_nb_for
 from collective.contact.plonegroup.utils import get_all_suffixes
 from collective.contact.plonegroup.utils import get_organizations
 from collective.contact.plonegroup.utils import get_own_organization
@@ -17,6 +20,7 @@ from imio.helpers.cache import invalidate_cachekey_volatile_for
 from imio.helpers.cache import setup_ram_cache
 from imio.helpers.content import get_modified_attrs
 from imio.helpers.content import richtextval
+from imio.helpers.content import safe_delattr
 from imio.helpers.security import fplog
 from imio.helpers.xhtml import storeImagesLocally
 from OFS.interfaces import IObjectWillBeAddedEvent
@@ -50,6 +54,7 @@ from Products.PloneMeeting.utils import AdviceAfterModifyEvent
 from Products.PloneMeeting.utils import AdviceAfterTransitionEvent
 from Products.PloneMeeting.utils import applyOnTransitionFieldTransform
 from Products.PloneMeeting.utils import get_annexes
+from Products.PloneMeeting.utils import get_internal_number
 from Products.PloneMeeting.utils import get_states_before
 from Products.PloneMeeting.utils import ItemAfterTransitionEvent
 from Products.PloneMeeting.utils import MeetingAfterTransitionEvent
@@ -644,6 +649,8 @@ def onItemCopied(item, event):
     item._update_predecessor(None)
     # remove link with Meeting
     item._update_meeting_link(None)
+    # remove internal_number
+    safe_delattr(item, "internal_number")
 
 
 def onItemMoved(item, event):
@@ -683,7 +690,8 @@ def item_added_or_initialized(item):
     # avoid multiple initialization
     # when using restapi for example, this empties adviceIndex
     # because init/update_local_roles/init
-    if hasattr(item, '_v_already_initialized'):
+    # wait for portal_type to be initialized
+    if item.portal_type == "MeetingItem" or hasattr(item, '_v_already_initialized'):
         return
     item._v_already_initialized = True
 
@@ -700,15 +708,15 @@ def item_added_or_initialized(item):
     item.deleted_children_history = PersistentList()
     # Add a place to store takenOverBy by review_state user id
     item.takenOverByInfos = PersistentMapping()
-    # An item has ben modified, use get_again for portlet_todo
-    invalidate_cachekey_volatile_for('Products.PloneMeeting.MeetingItem.modified', get_again=True)
     # if element is in a MeetingConfig, we mark it with IConfigElement interface
     if item.isDefinedInTool():
         alsoProvides(item, IConfigElement)
     else:
         noLongerProvides(item, IConfigElement)
-        # An item has ben modified
-        invalidate_cachekey_volatile_for('Products.PloneMeeting.MeetingItem.modified', get_again=True)
+        # Manage internal_number if activated in @@internalnumber-settings
+        get_internal_number(item, init=True)
+    # An item has ben modified
+    invalidate_cachekey_volatile_for('Products.PloneMeeting.MeetingItem.modified', get_again=True)
 
 
 def onItemInitialized(item, event):
@@ -1017,6 +1025,9 @@ def onItemEditCancelled(item, event):
        the _at_creation to True, it means we are creating an item from a template,
        we need to delete it if first edit was cancelled.'''
     if item._at_creation_flag and not item.isTemporary():
+        # rollback internal_number if used and defined
+        if _internal_number_is_used(item):
+            decrement_nb_for(item.portal_type)
         parent = item.getParentNode()
         parent.manage_delObjects(ids=[item.getId()])
 
@@ -1051,6 +1062,9 @@ def onItemWillBeRemoved(item, event):
     # If we are trying to remove the whole Plone Site or a MeetingConfig, bypass this hook.
     if event.object.meta_type in ['Plone Site', 'MeetingConfig']:
         return
+
+    # decrement internal_number if it was the last added item
+    decrement_if_last_nb(item)
 
     # update item predecessor and successors
     predecessor = item.get_predecessor()