+++ src/Products/PloneMeeting/safe_utils.py
@@ -17,6 +17,7 @@ from Products.PloneMeeting.utils import field_is_empty
 from Products.PloneMeeting.utils import fieldIsEmpty
 from Products.PloneMeeting.utils import get_annexes
 from Products.PloneMeeting.utils import get_gn_position_name
+from Products.PloneMeeting.utils import get_internal_number
 from Products.PloneMeeting.utils import get_next_meeting
 from Products.PloneMeeting.utils import get_prefixed_gn_position_name
 from Products.PloneMeeting.utils import get_public_url